requirements.txt
python-dotenv
discord-py
